+++ src/pptx/oxml/shapes/shared.py
@@ -315,6 +315,9 @@ class CT_NonVisualDrawingProps(BaseOxmlElement):
     hlinkHover: CT_Hyperlink | None = ZeroOrOne("a:hlinkHover", successors=_tag_seq[2:])
     id = RequiredAttribute("id", ST_DrawingElementId)
     name = RequiredAttribute("name", XsdString)
+    hidden: bool = OptionalAttribute(  # pyright: ignore[reportAssignmentType]
+        "hidden", XsdBoolean, default=False
+    )
     del _tag_seq